% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdInStock.R
\name{prdInStockUI}
\alias{prdInStockUI}
\title{产品分组UI}
\usage{
prdInStockUI(
  tabTitle = "回料粉碎记录表",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = prdInStockFunc_left,
  func_right = prdInStockFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品分组UI
}
\examples{
prdInStockUI()
}
