% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdPickup.R
\name{prdPickupUI}
\alias{prdPickupUI}
\title{产品分组UI}
\usage{
prdPickupUI(
  tabTitle = "回料使用记录表",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = prdPickupFunc_left,
  func_right = prdPickupFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品分组UI
}
\examples{
prdPickupUI()
}
